# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface LoginRequest {
  email: string;
  password: string;
}

export interface Caregiver {
  id: string;
  name: string;
  email: string;
  role: string;
  participantIds: string[];
}

export interface AuthLoginResponse {
  token: string;
  caregiver: Caregiver;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PipelineStage {
  label: string;
  count: number;
  color: string;
}

export interface ServiceRequest {
  id: string;
  participant: string;
  service: string;
  status: string;
  /** @nullable */
  provider: string | null;
  budget: string;
  urgency: string;
  date: string;
  ndisNumber: string;
}

export interface DashboardSummary {
  activeParticipants: number;
  openRequests: number;
  automationRate: number;
  matchedToday: number;
  pipeline: PipelineStage[];
  recentRequests: ServiceRequest[];
}

export interface Participant {
  id: string;
  name: string;
  ndisNumber: string;
  status: string;
  plan: string;
  budget: number;
  spent: number;
  age: number;
  primaryGoal: string;
  coordinator: string;
}

export interface Provider {
  id: string;
  name: string;
  abn: string;
  category: string;
  rating: number;
  reviews: number;
  status: string;
  location: string;
  participants: number;
  specialties: string[];
  responseTime: string;
}

export type GetRequestsParams = {
  status?: string;
};