# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: dashboard
    description: Dashboard summary
  - name: participants
    description: NDIS participants
  - name: requests
    description: Service requests
  - name: providers
    description: Service providers
paths:
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated caregiver
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current caregiver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caregiver"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary scoped to the authenticated caregiver
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary data filtered to the caregiver's assigned participants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /participants:
    get:
      operationId: getParticipants
      tags: [participants]
      summary: List participants assigned to the authenticated caregiver
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of participants assigned to the caregiver
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /participants/{id}:
    get:
      operationId: getParticipant
      tags: [participants]
      summary: Get a participant by ID (must be assigned to the caregiver)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied — participant not assigned to this caregiver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /requests:
    get:
      operationId: getRequests
      tags: [requests]
      summary: List service requests for the caregiver's assigned participants
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of service requests filtered to the caregiver's assigned participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a request by ID (must belong to an assigned participant)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /providers:
    get:
      operationId: getProviders
      tags: [providers]
      summary: List all providers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Get a provider by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Provider details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    Caregiver:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        participantIds:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - email
        - role
        - participantIds

    AuthLoginResponse:
      type: object
      properties:
        token:
          type: string
        caregiver:
          $ref: "#/components/schemas/Caregiver"
      required:
        - token
        - caregiver

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    DashboardSummary:
      type: object
      properties:
        activeParticipants:
          type: integer
        openRequests:
          type: integer
        automationRate:
          type: integer
        matchedToday:
          type: integer
        pipeline:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStage"
        recentRequests:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRequest"
      required:
        - activeParticipants
        - openRequests
        - automationRate
        - matchedToday
        - pipeline
        - recentRequests

    PipelineStage:
      type: object
      properties:
        label:
          type: string
        count:
          type: integer
        color:
          type: string
      required:
        - label
        - count
        - color

    Participant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ndisNumber:
          type: string
        status:
          type: string
        plan:
          type: string
        budget:
          type: number
        spent:
          type: number
        age:
          type: integer
        primaryGoal:
          type: string
        coordinator:
          type: string
      required:
        - id
        - name
        - ndisNumber
        - status
        - plan
        - budget
        - spent
        - age
        - primaryGoal
        - coordinator

    ServiceRequest:
      type: object
      properties:
        id:
          type: string
        participant:
          type: string
        service:
          type: string
        status:
          type: string
        provider:
          type: ["string", "null"]
        budget:
          type: string
        urgency:
          type: string
        date:
          type: string
        ndisNumber:
          type: string
      required:
        - id
        - participant
        - service
        - status
        - provider
        - budget
        - urgency
        - date
        - ndisNumber

    Provider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        abn:
          type: string
        category:
          type: string
        rating:
          type: number
        reviews:
          type: integer
        status:
          type: string
        location:
          type: string
        participants:
          type: integer
        specialties:
          type: array
          items:
            type: string
        responseTime:
          type: string
      required:
        - id
        - name
        - abn
        - category
        - rating
        - reviews
        - status
        - location
        - participants
        - specialties
        - responseTime