# lib/api-zod/src/generated/types/authLoginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Caregiver } from "./caregiver";

export interface AuthLoginResponse {
  token: string;
  caregiver: Caregiver;
}